(squashme) reorganize graph package
